Add Marty's compressed Golem
25% size, only 65f of loop, colours reduced to 32, took out every other frame and slowed animation to compensate. 
Done used Aspertite, ezgif allegedly can do the same stuff.

re: finding the loop
"I focus on a point where the animation hits the edge of the box, in this case its right hand. So I went to frame 41 where it touches the edge, then went to frame 106 where it does again, make sure they are actually the same exact image, and cut out everything before and after. Oh and I also delete one of the two of 41 and 106 so it's not duplicated at the beginning and end of the new loop" - Marty

2.34 MB -> 70 KB